Add direct children to Calculator component
