+++ .eslintrc.json
@@ -18,6 +18,7 @@
       "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }],
       "react/react-in-jsx-scope": "off",
       "import/no-unresolved": "off",
+      "import/no-extraneous-dependencies": "off",
       "no-shadow": "off"
     },
     "overrides": [
