+++ src/components/ButtonBox.js
@@ -1,8 +1,13 @@
-/* eslint-disable react/prop-types */
+import PropTypes from 'prop-types';
+
 const ButtonBox = ({ children }) => (
   <div className="buttonBox">
     {children}
   </div>
 );
+
+ButtonBox.propTypes = {
+  children: PropTypes.node.isRequired,
+};
 
 export default ButtonBox;
